Initial implementation of ceilometer-alarm
The implemented use case is raising an alarm (i.e. executing something)
as soon as a cumulative counter is above or below a certain threshold.

Alarms are stored in alarms.json inside the config directory.

The result of the "statistic" during a "period" is compared
to the alarm "threshold" with the operator in "comparaison_operator"
field of the alarm.

the statistic is calculated with get_meter_statistics() of the storage
connection.

The criteria handled to check if the alarm threshold needs to be checked
for a meter are for now only:
- counter_name
- resource_id

When a alarm is triggered a shell script is executed, same things when the
alarm is untriggered

The used shell script are stored in fields "ok_action" and "alarm_action" of the
alarm description.

The alarm status stays in memory as long as the alarms aren't stored in a
persistante storage.

Change-Id: I2dfebc30a47576e83906a6d4011ce5d52041e901